requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pymongo==4.0.1
python-dotenv==0.19.0
pydantic==1.8.2
boto3==1.26.137
google-generativeai==0.2.0
python-multipart==0.0.5